requirements.txt
boto3==1.40.30
botocore==1.40.30
certifi==2025.8.3
charset-normalizer==3.4.3
confluent-kafka==2.11.1
idna==3.10
jmespath==1.0.1
numpy==2.3.3
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
